[WIP] game start bug, I am instantly losing
